requirements.txt
# ===============================
# Footballer FaceGAN Requirements
# ===============================

# --- Core environment ---
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-image>=0.24.0
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.9.0
seaborn>=0.13.0

# --- GAN utilities & evaluation ---
torch-fidelity>=0.3.0        # FID, KID metrics
lpips>=0.1.4                 # perceptual similarity (optional)
einops>=0.7.0                # tensor rearrangements
scipy>=1.13.0

# --- Visualization & reporting ---
gradio>=4.37.0
plotly>=5.24.0
Pillow>=10.3.0
jupyterlab>=4.2.0
notebook>=7.2.0
markdown2>=2.5.0
tabulate>=0.9.0

# --- Development tools ---
black>=24.4.0
isort>=5.13.0
flake8>=7.0.0
autopep8>=2.2.0
pytest>=8.1.0

# --- Optional for latent-space analysis ---
scikit-learn>=1.5.0
umap-learn>=0.5.6

# --- Optional: report export ---
pypandoc>=1.13

# --- GPU/AMP utilities (if needed) ---
nvidia-ml-py>=12.535.108     # monitor GPU usage

# --- Notes ---
# To install: pip install -r requirements.txt
# To verify GPU: python -m torch.utils.collect_env
# Tested with Python 3.10–3.11
